# docs/openapi.generated.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/simulation/building-envelope/{parcelId}": {
      "get": {
        "operationId": "SimulationController_buildingEnvelope",
        "parameters": [
          {
            "name": "parcelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userReference",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "simulation"
        ]
      }
    },
    "/simulation/merge-candidates/{parcelId}": {
      "get": {
        "operationId": "SimulationController_mergeCandidates",
        "parameters": [
          {
            "name": "parcelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "simulation"
        ]
      }
    },
...
  },
  "info": {
    "title": "Türkiye E-İmar Platform API",
    "description": "Backend foundation for source discovery, connector orchestration, parcel queries, and GIS analysis.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}